chore: better install action code style
