--- infra/rush-x/src/actions/index.ts
@@ -1,1 +1,0 @@
-export { incrementActions } from './increment/actions';
